Add files via upload
This new version of RANKS produces a single score per gene which captures both sgRNA depletion (negative scores) and sgRNA enrichment (positive scores) instead of producing a separate depletion and enrichment score. It is important to run the new version of the control-distribution.pl script before using the new version of RANKS because the score distribution is different.